Refactored field_extrator to field_remover
